+++ test/td_dq_web/controllers/rule_result_controller_test.exs
@@ -131,6 +131,7 @@ defmodule TdDqWeb.RuleResultControllerTest do
                  "implementation_key" => "ri135",
                  "params" => %{"param3" => "5"},
                  "records" => 4,
+                 "result_type" => "percentage",
                  "result" => "0.00"
                },
                %{
@@ -139,6 +140,7 @@ defmodule TdDqWeb.RuleResultControllerTest do
                  "implementation_key" => "ri135",
                  "params" => %{"param1" => "valor", "param2" => "valor2", "param3" => "4"},
                  "records" => 1_000_000,
+                 "result_type" => "percentage",
                  "result" => "99.99"
                }
              ]